+++ .gitignore
@@ -1,1 +1,3 @@
 build/
+gmon.out
+out.prof
