+++ src/Systems.cpp
@@ -9,7 +9,6 @@ std::vector<std::function<void(Game*, double)>> getSystems() {
     systems.push_back(cameraFollowerSystem);
     systems.push_back(aiSystem);
     systems.push_back(healthSystem);
-    systems.push_back(meleeSystem);
     systems.push_back(projectileSystem);
     systems.push_back(playerSystem);
     systems.push_back(particleSystem);
