+++ src/data/Dungeon.hpp
@@ -43,7 +43,7 @@ class Dungeon {
         }
 
         bool rayCast(raylib::Vector2 start, raylib::Vector2 end, float* outDistance = nullptr);
-        std::vector<raylib::Vector2> pathFind(raylib::Vector2 start, raylib::Vector2 end, std::function<double(raylib::Vector2, raylib::Vector2, int)> heuristic);
+        std::vector<raylib::Vector2> pathFind(raylib::Vector2 start, raylib::Vector2 end);
         // TODO: Make a better pathfinding system.
 
         int width;
